refactor 로그아웃 로직 분리 및 헤더 리팩토링
